Genetic Algorithm v1.0.0 Gene

This struct is representing a single value in Chromosome.

Its key v is used for storing value of Gene

Example

  # Create new Gene with some value.
  %Gene{v: 1}

Types

t()

t() :: %Gene{v: any()}