Genetic Algorithm v1.0.0 GeneticAlgorithm

This module is grouping functions from GeneticAlgorithm.Helpers to easily find the solution of a given problem.

Options

Those are options with their default values:

%{
  gene_values: [0, 1],  # values used by Gene
  length: 5,            # length of a Chromosome
  chrom_in_gen: 1000,   # population size
  min_fitness: 0.9,     # minimum fitness allowing to stop
  mutation_chance: 0.1, # mutation chance
  shown_chrom: 3        # number of Chromosomes returned as solution
}

Types

fitness_f()

fitness_f() :: function(genes :: [Genes.t()]) :: float()

fitness_f describes fitness function which should be passed to various functions.

solution()

solution() :: %{generations: non_neg_integer(), most_fitting: [Chromosome.t()]}

solution describes solution returned after stopping the algoritm.

Functions

find_solution(fitness_func, opts \\ [])

find_solution(fitness_func :: fitness_f(), opts :: keyword()) :: solution()

Finds solution based on given fitness_func.

For list of available options see GeneticAlgorithm documentation.

generate_initial_population(fitness_func, opts \\ [])

generate_initial_population(fitness_func :: fitness_f(), opts :: keyword()) :: [
  Chromosome.t()
]

Generates new population with given opts.

opts:

:chrom_in_gen

:gene_values

:length

Also see GeneticAlgorithm.

next_generation(population, fitness_func, opts \\ [])

next_generation(
  population :: [Chromosome.t()],
  fitness_func :: fitness_f(),
  opts :: keyword()
) :: [Chromosome.t()]

Advances population of one generation.

opts:

:mutation_chance

:gene_values

:chrom_in_gen

Also see GeneticAlgorithm.