Genetic Algorithm v1.0.0 GeneticAlgorithm.Helpers

This module is grouping functions used to generate, crossover and mutate genetic algoritm chromosomes population.

Types

fitness_f()

fitness_f() :: function(genes :: [Genes.t()]) :: float()

Functions

crossover(population, fitness_func)

crossover(population :: [Chromosome.t()], fitness_func :: fitness_f()) :: [
  Chromosome.t()
]

Crossovers Chromosomes in given population.

Creates new population based on previous and having same amount of elements.

mutate(population, allowed_values, chance, fitness_func)

mutate(
  population :: [Chromosome.t()],
  allowed_values :: list(),
  chance :: float(),
  fitness_func :: fitness_f()
) :: [Chromosome.t()]

Mutates every Chromosome in population with given chance.

Example

iex> p = [%Chromosome{fitness: 1.0, genes: [%Gene{v: 0}]}]
iex> GeneticAlgorithm.Helpers.mutate(p, [0, 1], 1.0, fn _ -> 1.0 end)
[%Chromosome{fitness: 1.0, genes: [%Gene{v: 1}]}]

populate(size, allowed_values, length, fitness_func)

populate(
  size :: non_neg_integer(),
  allowed_values :: list(),
  length :: non_neg_integer(),
  fitness_func :: fitness_f()
) :: [Chromosome.t()]

Generates new random population with given parameters.

size - population size

allowed_values - values used as Gene values

length - amount of Genes in Chromosome

fitness_func - function calculating Chromosome fitness

Examples

iex> GeneticAlgorithm.Helpers.populate(2, [0], 1, fn _ -> 1.0 end)
[
  %Chromosome{fitness: 1.0, genes: [%Gene{v: 0}]},
  %Chromosome{fitness: 1.0, genes: [%Gene{v: 0}]}
]

iex> GeneticAlgorithm.Helpers.populate(0, [0], 1, fn _ -> 1.0 end)
[]

rulette_select(population, n)

rulette_select(population :: [Chromosome.t()], n :: non_neg_integer()) :: [
  Chromosome.t()
]

Selects n elements from a population according to their fitness.

If Chromosome has more fitness it has greater chance to be chosen.

select_most_fitting(population, n)

select_most_fitting(population :: [Chromosome.t()], n :: non_neg_integer()) :: [
  Chromosome.t()
]

Selects n most fitting Chromosomes from a population.

Chromosomes are sorted by fitness, ascending.